pyproject.toml
[project]
name = "z-image-turbo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers>=0.36.0",
    "python-slugify>=8.0.4",
    "sdnq>=0.1.3",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "httpx>=0.28.1",
    "numpy",
    "opencv-python",
    "basicsr>=1.4.2",
    "realesrgan",
]

[tool.uv.sources]
realesrgan = { path = "local_packages/realesrgan" }
